Rely on USERPROFILE instead of HOMEPATH (#37)
Processes started with alternate credentals using 'runas' functionality
(either with the runas command or with the 'Run as different user' menu
in explorer do not use the same HOMEDRIVE and HOMEPATH variables as
when the process is started by the logged-in user.
As such when starting Powershell using runas, the zlocation.txt ends
up in e.g. the current working directory when the process was started
(when starting shells as different user in Conemu for instance) or in
c:\windows\system32 (using runas on commandline).
The USERPROFILE environment variable does not have this problem and
always points to the same path so this is more robust.