Catch errors likely caused by null paths in db
The errors caused by reading the db when a location with no path set
has got into the database are actually `[InvalidCastException]`s thrown
when trying to convert the identifiers from the database into strings
for `[Location]`s.

This is caused by entries inserted with no path calling `GetById($null)`
so they get the 0-parameter version,
https://github.com/mbdavid/LiteDB/blob/f7ee275908ca48156a9d0bed054809955843e4b0/LiteDB/Document/BsonValue.cs#L49

Restricting the `catch{}` to `[IOException]` also means it's possible to
handle the file in use errors the retry logic is supposed to be
handling separately to other types of error.