Fix instructions regarding profile in README
Mention the directory needs to exist otherwise users are greeted with
rather meaningless 'The system cannot find the path specified.' errors.
Remove the instructions which hardcode the profile directory because
they are incorrect for PowerShell Core and teaching users to use
$PROFILE is better anyway.
Mention the default directory used for PowerShell Core as well to make
sure users know there are multiple locations.